Add support for shelly floods
This adds support for the Shelly Flood device, which is able to publish
to MQTT. This includes support for the four basic sensors that the Flood
provides. There needed to be a few changes to the way topics are handled
to accommodate them, but it should be more flexible for other devices
moving forward.

Specifically, Flood devices publish their statuses as single values
across multiple topics, as opposed to as a JSON object on one topic. To
accommodate that, the way that topics are mapped to devices had to be
modified so there could be multiple topics mapped to a single device.
Adjusted the other devices to match the new pattern.

Also centralized the place where we get the device address from the
device object; I noticed there was one other place where we weren't
generating the address the same way as we were on creation.